TypeScript migration and other miscellaneous fixes and upgrades
In this commit:
- full migration to TypeScript
- webpack build script in package.json to build the typescript depenancies
- argument change in robot.perform that aims to lead to the deprecation of the facing attribute

The previous commit was for safety purposes, and many of the changes have since been reverted.
